+++ lucid/models/weights/__init__.py
@@ -42,6 +42,7 @@ from lucid.models.vision.convnext._weights import (
     ConvNeXtSmallWeights,
     ConvNeXtBaseWeights,
     ConvNeXtLargeWeights,
+    ConvNeXtXLargeWeights,
 )
 
 __all__ = [
@@ -54,4 +55,5 @@ __all__ = [
     "ConvNeXtSmallWeights",
     "ConvNeXtBaseWeights",
     "ConvNeXtLargeWeights",
+    "ConvNeXtXLargeWeights",
 ]